+++ mobile/src/main/java/com/moysa/saveup/data/model/Transaction.java
@@ -12,7 +12,7 @@ public interface Transaction {
 
     int getPocketId();
 
-    int getAmount();
+    float getAmount();
 
     String getComment();
